Re-factor some warnings

+++ anvil/include/texturemanager.h
@@ -14,26 +14,26 @@ public:
     static TextureManager* instance();
     bool loadTexture(std::string filename, std::string id, SDL_Renderer* renderer);
     /* Draw the texture of given width and height */
-    void draw(std::string id, int x, int y, int width, int height,
+    void draw(std::string id, float x, float y, float width, float height,
               SDL_Renderer* renderer);
 
     /* Draw certain frame from the spreadsheet */
-    void drawFrame(std::string id, int x, int y, int width, int height,
+    void drawFrame(std::string id, float x, float y, float width, float height,
                    int currentRow, int currentFrame,
                    SDL_Renderer* renderer);
 
-    void drawFrameScaled(std::string id, float scale, int x, int y, int width, int height,
+    void drawFrameScaled(std::string id, float scale, float x, float y, float width, float height,
         int currentRow, int currentFrame,
         SDL_Renderer* renderer);
 
     void drawText(std::string id, std::string text,
                   TTF_Font* font, SDL_Color color,
-                  int x, int y, int width, int height,
+                  float x, float y, float width, float height,
                   SDL_Renderer* renderer);
 
     void drawTextWrapped(std::string id, Uint32 wrapLength, std::string text,
         TTF_Font* font, SDL_Color color,
-        int x, int y, int width, int height,
+        float x, float y, float width, float height,
         SDL_Renderer* renderer);
 
     void drawTextMultiline(SDL_Renderer* renderer,
@@ -54,10 +54,10 @@ public:
 
     std::pair<int, int> textureSize(std::string id);
 
-    void drawQuadrilateral(SDL_Renderer* renderer, int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4);
-    void drawPoint(SDL_Renderer *renderer, int x1, int y1);
-    void drawLine(SDL_Renderer* renderer, int x1, int y1, int x2, int y2);
-    void drawRect(SDL_Renderer* renderer, int x, int y, int w, int h, SDL_Color color);
+    void drawQuadrilateral(SDL_Renderer* renderer, float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4);
+    void drawPoint(SDL_Renderer *renderer, float x1, float y1);
+    void drawLine(SDL_Renderer* renderer, float x1, float y1, float x2, float y2);
+    void drawRect(SDL_Renderer* renderer, float x, float y, float w, float h, SDL_Color color);
 private:
     TextureManager();
     std::map<std::string, SDL_Texture*> m_texture_map;
